Introduce a Contact Sanitiser (#36)
Restore original email sanitiser regex to ensure that emails are in a
consistent format for sending signup invites to sponsees.

Previously the system was rejecting emails in the format of

  clay@example.com<mailto:clay@example.com>

Office365 was formatting email addresses in plain text form like
this.

This class will ensure emails are in a valid format along with phone
numbers.  It will also internationalise phone numbers.