Disable the auto confirmation behaviour for SNS
The auto confirmation behaviour was requried when setting
up the endpoint with SNS, but is no longer required as it
is now subscribed to SNS.

By leaving this functionality inside the app we allow other
SNS queues to subscribe to this endpoint, which we don't want
to allow.

Instead print out the notification to the logger, which allows
us to manually subscribe if ever required.