Extract seperate presenters for different stats
Volumetrics and CompletionRate are very similar when it comes
fetching data about user sign ups and they are exactly the
same when it comes to sending data to Performance service.

The only thing that differs is how data sent in payload is
structured. Extracting this logic to mutiple presenters allows
us to plug in whatever structure we need to send.