Silence Logger in test environment (#68)
We use Rake tasks which don't have access to the Rack logger, so we use
the logger from the Ruby standard library.

This is so we can have better visibility over what the rake tasks are
doing.

Unlike the Rack logger, we cannot switch this off in the test
environment with a configuration. Pull out the dependency on the
logger, and inject a fake during testing.