Use new email parser in SNS notification handler
The knowlege of the request is now being dealt with by the email parser.
We now process the entire request before handling it, which forces some
tests to change.
Use fetch instead of square brackets so we can raise a key error instead
of a nomethod error ([] on nil).  This can be rescued and we can short
curcuit execution.  At this point we will log to Cloudwatch.