Append whitespace to username for SMS notifications
Some devices do not present the login and password credentials
on separate lines. This leads to the username being concatenated
with the SMS template text preceding the password. This often
confuses users as they attempt to copy and paste these values into
their Wifi configuration.
This commit appends a single space char to the end of the login
value for clearer presentation in the SMS template.