Reset RACK_ENV in specs where it's overridden
In some tests, `ENV["RACK_ENV"]` is overridden to `staging` or `production`
either in a `before` block or inline.

ENV is not reset between tests, so any subsequent tests requiring `RACK_ENV`
to be `test` will then fail.

This change moves inlines to `before` blocks, and adds an `after` block
to reset back to `test`.