Handle Notifications::Client::BadRequestError
If a text message comes in from an invalid number, a Notifications::Client::BadRequestError
is raised. There are also other circumstances we haven't seen that could
trigger this, such as a text from a landline number.

Catch this error, and if it is due to a ValidationError (in the body of
the message) log it and rollback. Else throw the original error.

Either way we're now wrapping the user creation in a transaction, so on
either error the user will not be created in the database.