Introduce a custom error class to capture application specific errors
So that more errors are being sent to Sentry if something is to
go wrong. Any expected errors caused by malformed user content being
sent to the User signup API are caught using the UserSignupError class.

All other unspecified errors are now sent to Sentry to warn us about
a potential issue in the code / enviromnent